+++ class/actes/ActesTransactionsSQL.class.php
@@ -1,35 +1,32 @@
 <?php 
-class ActesTransactionsSQL {
+class ActesTransactionsSQL extends SQL{
 
-	public function __construct(SQLQuery $sqlQuery){
-		$this->sqlQuery = $sqlQuery;
-	}
-	
+
 	public function getInfo($id){
 		$sql = "SELECT * FROM actes_transactions WHERE id=?";
-		return $this->sqlQuery->queryOne($sql,$id);
+		return $this->queryOne($sql,$id);
 	}
 	
 	public function getStatusInfo($id,$status){
 		$sql = "SELECT * FROM actes_transactions_workflow WHERE transaction_id=? AND status_id=?";
-		return $this->sqlQuery->queryOne($sql,$id,$status);
+		return $this->queryOne($sql,$id,$status);
 	}
 
 	public function getLastStatusInfo($id){
 		$sql = "SELECT * FROM actes_transactions_workflow WHERE transaction_id=? ORDER BY date DESC LIMIT 1";
-		return $this->sqlQuery->queryOne($sql,$id);
+		return $this->queryOne($sql,$id);
 	}
 
 	
 	public function getAllFile($id){
 		$sql = "SELECT * FROM actes_included_files WHERE transaction_id=? ORDER BY id";
-		return $this->sqlQuery->query($sql,$id);
+		return $this->query($sql,$id);
 	}
 	
 	public function setArchiveURL($transaction_id,$archive_url){
 		$sql = "UPDATE actes_transactions SET archive_url=? " .
     			" WHERE id=?";
-		$this->sqlQuery->query($sql,$archive_url,$transaction_id);
+		$this->query($sql,$archive_url,$transaction_id);
 	}
 
 
@@ -39,44 +36,44 @@ class ActesTransactionsSQL {
 	    $sql = "INSERT INTO actes_transactions_workflow (transaction_id, status_id, date, message,flux_retour) " .
 	    		" VALUES( ? , ? , ? , ? ,?) RETURNING ID";
 
-  		$id = $this->sqlQuery->queryOne($sql,$transaction_id,$status_id,$date,$message,$flux_retour);
+  		$id = $this->queryOne($sql,$transaction_id,$status_id,$date,$message,$flux_retour);
 
 
     	$sql = "UPDATE actes_transactions SET last_status_id=? " .
     			" WHERE id=?";
 
-    	$this->sqlQuery->query($sql,$status_id,$transaction_id);
+    	$this->query($sql,$status_id,$transaction_id);
 		return $id;
 	}
 	
 	public function getArchiveFStatus($status_id){
 		$sql = "SELECT  actes_transactions.id as id FROM actes_transactions " .
 				" WHERE last_status_id=? ";
-		return $this->sqlQuery->query($sql,$status_id);
+		return $this->query($sql,$status_id);
 	}
 
 	public function getArchiveFromStatusWithSAE($status_id){
         $sql = "SELECT  *,actes_transactions.id as id FROM actes_transactions " .
             " JOIN authorities ON actes_transactions.authority_id=authorities.id " .
             " WHERE last_status_id=? AND authorities.pastell_url IS NOT NULL AND authorities.pastell_url != '' ";
-        return $this->sqlQuery->query($sql,$status_id);
+        return $this->query($sql,$status_id);
     }
 
     public function getEnveloppeIdByTransactionsStatus($last_status_id,$antivirus_check = true){
         $sql = "SELECT DISTINCT envelope_id FROM actes_transactions WHERE last_status_id=? AND antivirus_check=?";
-        return $this->sqlQuery->queryOneCol($sql,$last_status_id,$antivirus_check);
+        return $this->queryOneCol($sql,$last_status_id,$antivirus_check);
     }
 
     public function getIdByEnvelopeId($envelope_id){
         $sql = "SELECT id FROM actes_transactions WHERE envelope_id=?";
-        return $this->sqlQuery->queryOneCol($sql,$envelope_id);
+        return $this->queryOneCol($sql,$envelope_id);
 
     }
 
 	public function getTransactionIdFromStatus($status_id, $antivirus_check = true){
 		$sql = "SELECT  actes_transactions.id as id FROM actes_transactions " .
 				" WHERE last_status_id=? AND antivirus_check=?";
-		return $this->sqlQuery->queryOneCol($sql,$status_id,$antivirus_check);
+		return $this->queryOneCol($sql,$status_id,$antivirus_check);
 	}
 
     public function getLastArchiveFromStatus($status_id,$start_date){
@@ -86,7 +83,7 @@ class ActesTransactionsSQL {
             " AND actes_transactions_workflow.status_id=? ".
             " WHERE last_status_id=? AND authorities.pastell_url IS NOT NULL " .
             " AND authorities.pastell_url != '' AND actes_transactions_workflow.date > ?";
-        return $this->sqlQuery->query($sql,$status_id,$status_id,$start_date);
+        return $this->query($sql,$status_id,$status_id,$start_date);
     }
 
 	
@@ -95,13 +92,13 @@ class ActesTransactionsSQL {
 				" FROM actes_transactions " .
 				" JOIN actes_envelopes ON actes_transactions.envelope_id=actes_envelopes.id " .
 				" WHERE last_status_id=13 OR last_status_id = 6";
-		return $this->sqlQuery->query($sql);	
+		return $this->query($sql);
 	}
 	
 		
 	public function getNextNumeroTransfert(){
 		$sql = "SELECT count(*) + 1 FROM actes_transactions_workflow WHERE status_id=? AND date(date) = date(now());";
-		return $this->sqlQuery->queryOne($sql,12);
+		return $this->queryOne($sql,12);
 	}
 	
 	public function getLatestDate($id){
@@ -114,7 +111,7 @@ class ActesTransactionsSQL {
 				" FROM actes_transactions_workflow " .
 				" WHERE status_id IN (4,11,7) " .
 				" AND transaction_id IN (".implode(",",$all_id).")";
-		return $this->sqlQuery->queryOne($sql);
+		return $this->queryOne($sql);
 	}
 	
 	public function getRelatedTransaction($id){
@@ -122,7 +119,7 @@ class ActesTransactionsSQL {
 		$sql = "SELECT * ".
   				" FROM actes_transactions ".
   				" WHERE related_transaction_id=?";
-		foreach($this->sqlQuery->query($sql,$id) as $line){
+		foreach($this->query($sql,$id) as $line){
 			$result[] = $line;
 			$result = array_merge($result,$this->getRelatedTransaction($line['id']));
 		}
@@ -131,7 +128,7 @@ class ActesTransactionsSQL {
 	
 	public function setSAETransferIdentifier($id,$transfer_identifier){
 		$sql = "UPDATE actes_transactions SET sae_transfer_identifier=? WHERE id=?";
-		$this->sqlQuery->query($sql,$transfer_identifier,$id);
+		$this->query($sql,$transfer_identifier,$id);
 	}
 	
 	public function getDateTampon($id){
@@ -140,41 +137,41 @@ class ActesTransactionsSQL {
 				" AND actes_transactions.id = ? ".
 				" AND actes_transactions_workflow.transaction_id = actes_transactions.id " .
 				" AND actes_transactions_workflow.status_id=?";
-		return $this->sqlQuery->queryOne($sql,$id,4);
+		return $this->queryOne($sql,$id,4);
 	}
 	
 	public function getTransactionForAntiVirus(){
 		$sql = "SELECT DISTINCT id FROM actes_transactions WHERE last_status_id=? AND antivirus_check=?";
-		return $this->sqlQuery->queryOneCol($sql,1,0);
+		return $this->queryOneCol($sql,1,0);
 	}
 	
 	public function setAntivirusCheck($transaction_id){
 		$sql = "UPDATE actes_transactions SET antivirus_check=? WHERE id=?";
-		$this->sqlQuery->query($sql,1,$transaction_id);
+		$this->query($sql,1,$transaction_id);
 	}
 	
 	public function updateLastStatusId(){
 		$sql = "SELECT id FROM actes_transactions";
 	
-		$id_list = $this->sqlQuery->queryOneCol($sql);
+		$id_list = $this->queryOneCol($sql);
 		foreach($id_list as $transaction_id) {
 			//echo "$transaction_id : ";
 			$status=$this->getlaststatusforid($transaction_id);
 			//echo "$status\n";
 			$sqlupdate = "update actes_transactions set last_status_id = ? where  id = ?";
-			$this->sqlQuery->query($sql,$status,$transaction_id);
+			$this->query($sql,$status,$transaction_id);
         }
     }
 	
 	public function getlaststatusforid($idtrans){
 		$sql = " select status_id from actes_transactions_workflow where transaction_id = ? ORDER BY id DESC limit 1";
-		$status = $this->sqlQuery->queryOneCol($sql,$idtrans);
+		$status = $this->queryOneCol($sql,$idtrans);
         return $status[0];
     }
 
     public function getLastTransactionWorkflowInfo($id){
         $sql = " select * from actes_transactions_workflow where transaction_id = ? ORDER BY id DESC limit 1";
-        return $this->sqlQuery->queryOne($sql,$id);
+        return $this->queryOne($sql,$id);
     }
 
     public function getNbTransactionByMonth(){
@@ -183,14 +180,19 @@ class ActesTransactionsSQL {
 			" WHERE actes_transactions.type = '1' AND submission_date>'2015-01-01'" .
 			" GROUP BY month" .
 			" ORDER BY month DESC";
-		return $this->sqlQuery->query($sql);
+		return $this->query($sql);
 	}
 
 	public function getBySirenAndNumeroInterne($siren,$numero_interne){
         $sql = "SELECT actes_transactions.id from actes_transactions " .
             " JOIN actes_envelopes ON actes_transactions.envelope_id = actes_envelopes.id " .
             " WHERE siren=? AND number=? AND type='1'";
-        return $this->sqlQuery->queryOne($sql,$siren,$numero_interne);
+        return $this->queryOne($sql,$siren,$numero_interne);
+    }
+
+    public function getNbByStatus($status_id){
+        $sql = "SELECT count(*) FROM actes_transactions WHERE last_status_id=?";
+        return $this->queryOne($sql,$status_id);
     }
 	
 }