+++ public.ssl/admin/index.php
@@ -9,29 +9,44 @@ if ($userInfo['role'] != 'SADM'){
 }
 
 
-$info = array(
+$helios_status = array(
 	1=>"Posté",
 	7 => "En traitement",
 	2=>"En attente de transmission",
 	3=>"Transmis"
 );
 
-/** @var HeliosTransactionsSQL $heliosTransactionsSQL */
-$heliosTransactionsSQL = $objectInstancier->{'HeliosTransactionsSQL'};
+$heliosTransactionsSQL = $objectInstancier->get('HeliosTransactionsSQL');
 
-$result = array();
-foreach($info as $status_id => $status_libelle){
-	$result[$status_id] =  $heliosTransactionsSQL->getNbByStatus($status_id);
+$helios_nb_transaction_by_status = array();
+foreach($helios_status as $status_id => $status_libelle){
+	$helios_nb_transaction_by_status[$status_id] =  $heliosTransactionsSQL->getNbByStatus($status_id);
 }
 
 $today = date("Y-m-d");
 
 $nb_transaction_transmise_hier = $heliosTransactionsSQL->getNbByStatusAndDate(3,$today);
 
-
 $heliosResponsesError = new HeliosResponsesError();
+$helios_nb_responses_error = $heliosResponsesError->getNbError();
 
-$nb_responses_error = $heliosResponsesError->getNbError();
+
+$actes_status = array(
+    1=>"Posté",
+    2=>"En attente de transmission",
+    3=>"Transmis"
+);
+$actesTransactionsSQL = $objectInstancier->get('ActesTransactionsSQL');
+
+
+$actes_nb_transaction_by_status = array();
+foreach($actes_status as $status_id => $status_libelle){
+    $actes_nb_transaction_by_status[$status_id] =  $actesTransactionsSQL->getNbByStatus($status_id);
+}
+
+$actesResponsesError = $objectInstancier->get('ActesResponsesError');
+$actes_nb_responses_error = $actesResponsesError->getNbError();
+
 
 
 $menuHTML = new MenuHTML();
@@ -60,6 +75,37 @@ ob_start();
         <a href="/admin/message/index.php">Publier un message d'urgence</a>
 
 
+        <h2>Actes : Nombre de transactions en cours</h2>
+        <table  class="data-table table table-striped ">
+            <tr>
+                <th>Type</th>
+                <th>Nombre de transactions</th>
+                <th>&nbsp;</th>
+            </tr>
+            <?php foreach($actes_nb_transaction_by_status as $status_id => $nb): ?>
+                <tr>
+                    <td><?php echo $actes_status[$status_id]?></td>
+                    <td><?php echo $nb?></td>
+                    <td>
+
+                        <a href="/modules/actes/index.php?status=<?php echo $status_id ?>" class="icon">
+                            Liste
+                        </a>
+                    </td>
+                </tr>
+            <?php endforeach ?>
+            <tr class="<?php echo $actes_nb_responses_error?"danger":"success" ?>">
+                <td>Emails reçus depuis Actes en erreur</td>
+                <td><span class="label label-<?php echo $actes_nb_responses_error?"danger":"success" ?>"><?php echo $actes_nb_responses_error ?></span></td>
+                <td>
+                    <a href="/modules/actes/admin/responses-actes-error.php" >
+                        Liste
+                    </a>
+                </td>
+            </tr>
+
+        </table>
+
 		<h2>Helios : Nombre de transactions en cours</h2>
 		<table  class="data-table table table-striped ">
 			<tr>
@@ -67,9 +113,9 @@ ob_start();
 				<th>Nombre de transactions</th>
 				<th>&nbsp;</th>
 			</tr>
-			<?php foreach($result as $status_id => $nb): ?>
+			<?php foreach($helios_nb_transaction_by_status as $status_id => $nb): ?>
 				<tr>
-					<td><?php echo $info[$status_id]?></td>
+					<td><?php echo $helios_status[$status_id]?></td>
 					<td><?php echo $nb?></td>
 					<td>
 
@@ -88,9 +134,9 @@ ob_start();
 					</a>
 				</td>
 			</tr>
-			<tr class="<?php echo $nb_responses_error?"danger":"success" ?>">
+			<tr class="<?php echo $helios_nb_responses_error?"danger":"success" ?>">
 				<td>Fichiers reçus depuis Helios en erreur</td>
-				<td><span class="label label-<?php echo $nb_responses_error?"danger":"success" ?>"><?php echo $nb_responses_error ?></span></td>
+				<td><span class="label label-<?php echo $helios_nb_responses_error?"danger":"success" ?>"><?php echo $helios_nb_responses_error ?></span></td>
 				<td>
 					<a href="/modules/helios/admin/responses-helios-error.php" >
 						Liste